pyproject.toml
[tool.poetry]
name = "webdavplay"
version = "0.1.0"
description = ""
authors = ["Kamaal <kamaal.f1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
